generate_nonce_with_grinding function in grinding.rs file
